requirements.txt
pytest>=6.2.4
evdev>=1.4.0
pyzmq>=22.1.0
